% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jags.R
\name{assign_iterations_mcarray}
\alias{assign_iterations_mcarray}
\title{Assign iterations mcarray}
\usage{
assign_iterations_mcarray(x, value, at)
}
\arguments{
\item{x}{mcarray object.}

\item{value}{mcarray object with same dimension except for the penultimate
dimension (iterations).}

\item{at}{vector of integer values at which to place the \code{value} array.}
}
\description{
Assign values to an mcarray object. Note this is not straightforward because
for a scalar one would use \code{x[, 1:10, ] <- value}, whereas for a matrix
one would use \code{x[, , 1:10, ] <- value}.
}
\examples{
\dontrun{
  x <- array(seq(180), dim = c(2, 2, 15, 3))
  value <- 1000 * x[, , 6:10, ]
  at <- 6:10
  assign_iterations_mcarray(x, value, at)
}

}
