% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chol-solve.R
\name{chol_solve}
\alias{chol_solve}
\title{Invert Symmetric Positive Definite Matrix via Cholesky Decomposition}
\usage{
chol_solve(x, tol = 100 * .Machine$double.eps)
}
\arguments{
\item{x}{matrix. symmetric (up to \code{tol}) and positive definite.}

\item{tol}{numeric. Differences of symmetry smaller than this value
are not considered significant.}
}
\value{
the matrix inverse of \code{x}.
}
\description{
Consider a matrix that is symmetric and has positive eigenvalues, and is
therefore postive definite. Then finds the inverse of the cholesky
decomposition of this matrix as a more numerically stable alternative
to \code{base::solve(x)}.
}
