% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibtex.R
\name{collate_bibtex}
\alias{collate_bibtex}
\title{Collate BibTeX Databases}
\usage{
collate_bibtex(path = ".")
}
\arguments{
\item{path}{character. The directory to search recursively for .bib files.}
}
\value{
\code{\link[RefManageR]{BibEntry}}.
}
\description{
Simple function that will find all .bib files in a directory and combine them
into a single \code{BibEntry} object.
}
\seealso{
\code{\link[RefManageR]{ReadBib}}.
}
