% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff-create-coeffs.R
\name{ff_create_coeffs}
\alias{ff_create_coeffs}
\title{Create Coefficients (Function Factory)}
\usage{
ff_create_coeffs(size, lim = 20, pm = TRUE, std_dev = 0.1)
}
\arguments{
\item{size}{numeric. Number of coefficients in each vector.}

\item{lim}{numeric. The range at which coefficients are initially set.}

\item{pm}{(plus / minus) logical. When TRUE coefficients can be
initially set to \eqn{\[-lim, lim\]}.}

\item{std_dev}{numeric. The SD of the nugget term added to each coefficient.}
}
\value{
A call such as \code{x <- ff_create_coeffs(...)} would allow \code{x} to be a
callable object that returns coefficients that can later be accessed using
\code{rlang::env_get(environment(x), "c_all")}.
}
\description{
Will return a function that is capable of returning different coefficients
each time while storing the generated coefficients in an object stored
in the function environment.
}
\details{
For more details on function factories
see (adv-r.hadley.nz/function-factories.html)\link{A}
}
