% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{format_p_values}
\alias{format_p_values}
\title{Format P Values with suffix}
\usage{
format_p_values(x, ...)
}
\arguments{
\item{x}{numeric. Values to be converted.}

\item{...}{extra arguments passed to \code{\link[base]{format}}.}
}
\description{
Convert p-values to an interpretable character vector with suffix shown in
details.
}
\details{
Suffix based on widely used:
"\code{Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1}"
}
\seealso{
\code{\link[base]{format.pval}} (no suffix).
}
