% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jags.R
\name{jags_samples_in_knitr}
\alias{jags_samples_in_knitr}
\title{Knitr Friendly Version of \code{jags.samples()}}
\usage{
jags_samples_in_knitr(
  model,
  variable.names,
  n.iter,
  thin = 1,
  type = "trace",
  force.list = FALSE,
  ...
)
}
\arguments{
\item{model}{a jags model object}

\item{variable.names}{a character vector giving the names of variables
    to be monitored}

\item{n.iter}{number of iterations to monitor}

\item{thin}{thinning interval for monitors}

\item{type}{type of monitor (can be vectorised)}

\item{force.list}{option to consistently return a named list of monitor
             types even if a single monitor type is requested}

\item{...}{optional arguments passed to the update method for jags
             model objects}
}
\description{
Modified version of \code{\link[rjags]{jags.samples}} that incorporates
[https://rmflight.github.io/knitrProgressBar/]{knitrProgressBar}.
}
\details{
In the source code of \code{\link[rjags]{jags.samples}} and the
\code{\link[rjags]{update}}, a progress bar is called only if
\code{interactive()} returns \code{TRUE}.

This is not the case for Rmarkdown documents among other scenarios, hence
we modify the source code to display a different progress bar.
}
\seealso{
\code{\link[rjags]{jags.samples}} and its source code.
}
