% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jags.R
\name{jags_samples_quarto}
\alias{jags_samples_quarto}
\title{Notebook Friendly Version of \code{jags.samples()}}
\usage{
jags_samples_quarto(...)
}
\arguments{
\item{...}{passed to \code{\link[rjags]{jags.samples}}.}
}
\description{
Modified version of \code{\link[rjags]{jags.samples}} that incorporates
[https://rmflight.github.io/knitrProgressBar/]{knitrProgressBar}.
}
