% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jags.R
\name{jags_samples_with_pb}
\alias{jags_samples_with_pb}
\title{JAGS Samples with Progress Bar}
\usage{
jags_samples_with_pb(model, variable.names, n.iter, n.update, thin)
}
\arguments{
\item{model}{a jags model object}

\item{variable.names}{a character vector giving the names of variables
    to be monitored}

\item{n.iter}{number of iterations to monitor}

\item{n.update}{integer. Number of (unsaved) warn-up iterations.}

\item{thin}{thinning interval for monitors}
}
\description{
\code{\link[rjags]{jags.samples}} default progress bar will not work when
\code{interactive()} returns \code{FALSE}. This method uses the \code{knitrProgressBar}
package with many \code{jags.samples()} calls.
}
