% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr-hooks.R
\name{knit-hook-chunk}
\alias{knit-hook-chunk}
\alias{knit_hook_summary}
\title{Knit (Chunk) Hooks}
\usage{
knitr::knit_hooks$set(NAME = knit_hook_*)

knit_hook_summary(before, options, envir)
}
\arguments{
\item{before, options, envir}{required arguments.}
}
\description{
Knit (chunk) hooks allow code to be ran before and after the execution of
code chunks, see \url{https://yihui.org/knitr/hooks} for more information.
}
\section{Functions}{
\itemize{
\item \code{knit_hook_summary}: Hide content inside of \code{<details><summary> ...} where the summary text
is the value of the chunk option.
}}

