% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr-tabset.R
\name{knit_tabset}
\alias{knit_tabset}
\title{List Tabsets (via \code{knit_print})}
\usage{
knit_tabset(x, ...)
}
\arguments{
\item{x}{list. Elements passed to \code{knit_print} within tabs,
names inform headings.}

\item{...}{extra arguments not used.}
}
\description{
To be used as a method for \code{knit_print} generic.
}
\seealso{
\code{vignette("knit_print", package = "knitr")}.
}
