% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr-setup.R
\name{knitr_default_setup}
\alias{knitr_default_setup}
\title{Knitr Defaults}
\usage{
knitr_default_setup(..., base_size = 18)
}
\arguments{
\item{...}{named arguments. Override knitr options.}

\item{base_size}{numeric. base font size for \code{ggplot2}, given in pts.}
}
\description{
Set (custom) default ggplot2 theme and knitr options.
}
