% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr-utilks.R
\name{knitr_print_tabset}
\alias{knitr_print_tabset}
\title{Print List Elements in tabset}
\usage{
knitr_print_tabset(x, level)
}
\arguments{
\item{x}{named list of objects to be printed.}

\item{level}{integer. How many \verb{#} should precede each heading title.}
}
\description{
Print elements in a list surrounded by defined headings, useful for
putting content into tabset within compataible \code{rmarkdown} documents.
}
