% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix-to-latex.R
\name{matrix_to_latex}
\alias{matrix_to_latex}
\title{Convert Matrix Object to LaTeX Code}
\usage{
matrix_to_latex(x, envir = "pmatrix")
}
\arguments{
\item{x}{object to be converted.}

\item{envir}{character. The environment the matrix will be surrounded by.
Defaults to \verb{\\\\begin\{pmatrix\}  \\\\end\{pmatrix\}}.}
}
\value{
a character is returned that produces the intended formatting when
it is passed to \code{cat()} or \code{writeLines()}.
}
\description{
Will take a matrix (numeric or character) and produce code that can be
placed into a .TeX document.
}
