% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jags.R
\name{parse.varnames}
\alias{parse.varnames}
\title{Parse Varnames}
\usage{
parse.varnames(varnames)
}
\description{
Internal \code{rjags} function required for \code{\link{jags_samples_in_knitr}}.
}
\keyword{internal}
