% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr-printers.R
\name{printer_tabset}
\alias{printer_tabset}
\title{Print Tabsets (via \code{knit_print})}
\usage{
printer_tabset(x, options, ...)
}
\arguments{
\item{x}{list. Each element is passed to \code{knit_print} within tabs.}

\item{options, ...}{arguments required for a \code{\link[knitr:knit_print]{knit_print}}
method.}
}
\description{
Converts a list to a Quarto / RMD tabset using the names as tab headings.
}
\section{Reporting}{

Either define a method, say \verb{knit_print.foo <- printer_*}, in the report
or set the \code{render} chunk option equal to the desired function.
}

\seealso{
\itemize{
\item Generic documentation, \code{\link[knitr:knit_print]{knit_print}}.
\item Detailed vignette, \code{vignette("knit_print", package = "knitr")}.
\item Related \code{\link[printr:printr]{printr}} package.
}

Other Knitr printer methods: 
\code{\link{printer_equation}()}
}
\concept{Knitr printer methods}
