% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto-render-here.R
\name{quarto_render_here}
\alias{quarto_render_here}
\title{Quarto Render}
\usage{
quarto_render_here(input, output_dir, ..., with_yaml = TRUE)
}
\arguments{
\item{input}{path. Quarto input file.}

\item{output_dir}{directory. Target directory to render target in.}

\item{...}{passed to \code{\link[quarto:quarto_render]{quarto::quarto_render()}}.}

\item{with_yaml}{logical. Should all \code{.yml} files also be used.}
}
\description{
Copies the chosen quarto file and YAML and renders within that directory.
}
