% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{range_expander}
\alias{range_expander}
\title{Range Expander}
\usage{
range_expander(..., scaling)
}
\arguments{
\item{...}{passed to \code{\link[base]{range}}.}

\item{scaling}{numeric. Scaling factor, will stretch when greater than 1
and shrink when less than 1.}
}
\value{
numeric of length 2. Consider \code{x = 0:10}, (spans 10 units),
\itemize{
\item \code{range_expander(x, scaling = 0.5)} returns \code{c(2.5, 7.5)} (5 units),
\item \code{range_expander(x, scaling = 1.5)} returns \code{c(-5, 15)} (20 units).
}
}
\description{
Modifies the output of \code{\link[base]{range}} to stretch or shrink.
This is mainly useful for setting plot limits.
}
\section{Future Development}{

After testing, this function will likely be moved to \code{jcutils} package.
}

