% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repr-pvals.R
\name{repr_pvals}
\alias{repr_pvals}
\title{Representation of p-values}
\usage{
repr_pvals(pv, digits = 3)
}
\arguments{
\item{pv}{numeric vector. P-values to be converted.}

\item{digits}{integer indicating the number of decimal places
    (\code{round}) or significant digits (\code{signif}) to be used.
    Negative values are allowed (see \sQuote{Details}).}
}
\description{
Turns numeric values denoting p-values to characters showing a more
informative representation of significance
}
\details{
Based on widely used:
"Signif. codes:  0 ‘\emph{\strong{’ 0.001 ‘}’ 0.01 ‘}’ 0.05 ‘.’ 0.1 ‘ ’ 1"
}
