% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_renderer.R
\name{rmd_renderer}
\alias{rmd_renderer}
\title{R Markdown Renderer}
\usage{
rmd_renderer(name)
}
\arguments{
\item{name}{character scalar. This defines the report name that is required
for consistency. See File Structure for details.}
}
\value{
A function that will render reports in the \verb{output/} directory.
The parameters of this sub-function include \code{subname} and \code{...}, the latter
of which will be passed to \code{rmarkdown::render}.
}
\description{
Given some report name, will create a function that calls \code{rmarkdown::render}
with consistent input and output arguments. This allows for the same report
to be compiled multiple times with slightly different arguments.
}
\section{File Structure}{

All reports must be placed in their own folder (of the same name) within
"inst/reports". For instance, if a report called "example" is created
and saved as \code{inst/reports/example/example.Rmd} as below, the output
folder (also shown) can be created by running the code in the examples
section.\preformatted{  +-- inst/
  |   +-- reports/
  |   |   +-- example/
  |   |   |   +-- example.Rmd
  |   |   |   +-- output/
  |   |   |   |   +-- example-subname1.pdf
  |   |   |   |   +-- example-subname2.pdf
  |   |   |   |   +-- example_subname3.html
}
}

\examples{
\dontrun{
example_func <- rmd_renderer("example")
example_func(subname = "subname1")
example_func(subname = "subname2", params = list(seed = 2))
example_func(subname = "subname1", output_format = "html_document")
}

}
