% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd-helpers.R
\name{rmd_summary}
\alias{rmd_summary}
\alias{rmd_summary.default}
\alias{rmd_summary.mlm}
\title{Nice Model Summary in R Markdown}
\usage{
rmd_summary(object, fmt = c("html", "latex"), digits = 3, resp)

\method{rmd_summary}{default}(object, fmt = c("html", "latex"), digits = 3, resp)

\method{rmd_summary}{mlm}(object, fmt = c("html", "latex"), digits = 3, resp)
}
\arguments{
\item{object}{Typical fit object such as "lm", "mlm", "glm", etc.}

\item{fmt}{choice. Is the table to be displayed via HTML ("html")
or via PDF ("latex").}

\item{digits}{numeric. Significant figures to round to for all numerics
other than p values.}

\item{resp}{regular expression. Should match a single response variable.
Will be silently ignored for non-mlm objects.}
}
\description{
A better looking summary table with reference columns, rounding, p-value
highlighting that can be called directly in RMD.
}
