% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd-helpers.R
\name{rmd_table}
\alias{rmd_table}
\title{Nice Table in R Markdown.}
\usage{
rmd_table(x, fmt = c("html", "latex"), scroll_vert = NULL, ...)
}
\arguments{
\item{x}{data.frame. Data to be converted into a table.}

\item{fmt}{choice. Is the table to be displayed via HTML ("html")
or via PDF ("latex").}

\item{scroll_vert}{numeric. If the format is HTML, create a scroll box of this
height (in pixels)}

\item{...}{extra arguments to be passed to either \code{kableExtra::kbl}
or \code{kableExtra::kable_styling}}
}
\value{
Either a HTML table or LaTeX table, depending on format.
}
\description{
This function utilises the flexibility of kableExtra to choose some 'nice'
default tables for either HTML or PDF via LaTeX. (subjective)
}
\section{Future Development}{

\itemize{
\item Add in some method to alter column widths, likely as a ncol-length simplex.
}
}

