% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session-info.R
\name{session_info}
\alias{session_info}
\title{Output Session Information}
\usage{
session_info(git_repo = rprojroot::find_root("DESCRIPTION"))
}
\arguments{
\item{git_repo}{character. Path to be passed to
\code{\link[git2r]{repository_head}}.}
}
\description{
A \code{knitr}-friendly output of \code{\link[utils]{sessionInfo}} with added git
information provided by \code{git2r}.
}
