% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{to_latex}
\alias{to_latex}
\alias{to_latex.matrix}
\alias{to_latex.numeric}
\title{R Object to Latex}
\usage{
to_latex(x, ...)

\method{to_latex}{matrix}(x, digits = NULL, envir = "pmatrix", ...)

\method{to_latex}{numeric}(x, digits = NULL, envir = "pmatrix", ...)
}
\arguments{
\item{x}{object used to choose a method.}

\item{...}{extra arguments passed to method.}

\item{digits}{integer. Passed to \code{\link[=format]{format()}}.}

\item{envir}{character. LaTeX matrix style.}
}
\description{
Generic method that accepts some R object and converts it to LaTeX.
}
