% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-imports.R
\name{utils-imports}
\alias{utils-imports}
\title{Imports}
\description{
Imports from other packages used within this package
}
\section{Links}{

\itemize{
\item \code{\link[rlang:dot-data]{rlang::dot-data()}}, data pronoun.
\item \code{\link[rlang:op-null-default]{rlang::op-null-default()}}, default value for NULL operator.
\item \code{\link[magrittr:pipe]{magrittr::\%>\%()}}, magrittr pipe.
}
}

\keyword{internal}
