% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{wrap_inside_latex_environment}
\alias{wrap_inside_latex_environment}
\title{Wrap LaTeX code Inside LaTeX Environment}
\usage{
wrap_inside_latex_environment(x, envir, indent = 4)
}
\arguments{
\item{x}{object to be converted.}

\item{envir}{character. LaTeX environment name.}

\item{indent}{integer. How many spaces to be added before each newline.}
}
\value{
character that produces the intended formatting when passed
to \code{\link[base]{cat}} or \code{\link[knitr]{asis_output}}.
}
\description{
Wrap LaTeX code, stored as a character in R, inside an environment such as
\verb{\\begin\{envir\} ... \\end\{envir\}}.
}
