% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibtex.R
\name{collate_bibtex}
\alias{collate_bibtex}
\title{Collate BibTeX Entries}
\usage{
collate_bibtex(path, file = "references.bib", overwrite = FALSE)
}
\arguments{
\item{path}{character. The directory (and sub-directories) to search
for .bib files}

\item{file}{output Bibtex file.}

\item{overwrite}{logical. If TRUE, replaces 'file', if it already exists.}
}
\value{
(invisible) An object of class "bibentry",
containing all found references.
}
\description{
Simple function that will find all .bib files in a directory and combine them
into a single .bib file.
}
