% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr-hooks.R
\name{knit-hook-chunk}
\alias{knit-hook-chunk}
\alias{knit_hook_summary}
\title{Knit (Chunk) Hooks}
\usage{
knit_hook_summary(before, options, envir)
}
\arguments{
\item{before, options, envir}{required arguments.}
}
\description{
Knit (chunk) hooks to be ran before and after the execution of code chunks.
}
\section{Functions}{
\itemize{
\item \code{knit_hook_summary()}: Hide content inside of \verb{<details><summary> ...} where the summary text
is the value of the chunk option.

}}
\section{Knitr}{

To use any of these functions. use the following in a R markdown document:

\if{html}{\out{<div class="sourceCode">}}\preformatted{knitr::knit_hooks$set(NAME = knit_hook_*)
}\if{html}{\out{</div>}}

See \url{https://yihui.org/knitr/hooks} for more information.
}

