% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr-tabset.R
\name{knitr_tabset}
\alias{knitr_tabset}
\alias{knitr_print_tabset}
\title{Output List Elements as Tabset}
\usage{
knitr_tabset(.x, .f, type = c("quarto", "rmd"), ...)

knitr_print_tabset(x, level)
}
\arguments{
\item{.x}{list to be looped over, names inform tab headings.}

\item{.f}{function or formula. See \code{\link[purrr:as_mapper]{purrr::as_mapper()}}.}

\item{type}{choice, Quarto or R Markdown. Which rendering program is used?}

\item{...}{extra arguments passed to \code{.f}}

\item{x, level}{Deprecated.}
}
\value{
\code{.x} is silently returned to allow for piping.
}
\description{
Map elements through \code{.f} and wrap each output into a tab within a tabset.
}
\section{Functions}{
\itemize{
\item \code{knitr_print_tabset}: Deprecated, use \code{knitr_tabset(x, print, "rmd")}.
}}

