% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr-html.R
\name{list_to_html_tabset}
\alias{list_to_html_tabset}
\title{List to HTML tabset}
\usage{
list_to_html_tabset(x, level, id = "id")
}
\arguments{
\item{x}{named list.}

\item{level}{integer. How many \verb{#} should precede each heading title.}

\item{id}{character. String to be used in chunk labelling, useful to avoid
duplicate chunk labels.}
}
\description{
Evaluates elements of a named list using names add heading set as some
chosen level (for use in \code{rmarkdown} documents).
}
