% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{matrix_to_latex}
\alias{matrix_to_latex}
\title{Convert Matrix Object to LaTeX}
\usage{
matrix_to_latex(x, digits, envir = "pmatrix")
}
\arguments{
\item{x}{object to be converted.}

\item{digits}{numeric. Rounding value for better value display.}

\item{envir}{character. LaTeX matrix style,
see \url{https://www.overleaf.com/learn/latex/Matrices}.}
}
\value{
character that produces the intended formatting when passed
to \code{\link[base]{cat}} or \code{\link[knitr]{asis_output}}.
}
\description{
Convert matrix to a (LaTeX) string to be printed "asis".
}
