% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr-printers.R
\name{printer_equation}
\alias{printer_equation}
\title{Print Equations (via \code{knit_print})}
\usage{
printer_equation(x, options, ...)
}
\arguments{
\item{x}{object to be passed to \code{\link{to_latex}}.}

\item{options, ...}{arguments required for a \code{\link[knitr:knit_print]{knit_print}}
method.}
}
\description{
Converts a base R object into LaTeX output within Quarto / RMD.
}
\section{Reporting}{

These methods are to be used in Quarto / RMD reports.

For global usage on all objects of class \code{foo}, register as an S3 method,
typically in the setup chunk.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{library(knitr)
registerS3method("knit_print", "foo", printer_*)
}\if{html}{\out{</div>}}

For one-off usage, set the \code{render} chunk option, that is:

\if{html}{\out{<div class="sourceCode markdown">}}\preformatted{```\{r\}
#| render: !expr printer_*

```
}\if{html}{\out{</div>}}
}

\section{Chunk Options}{

\itemize{
\item \code{eqn_name} character. Name to place on the LHS of the equation.
\item \code{eqn_digits} integer. Passed to \code{\link{to_latex}}.
\item \code{eqn_envir} character. Passed to \code{\link{to_latex}}.
}
}

\seealso{
\itemize{
\item Generic documentation, \code{\link[knitr:knit_print]{knit_print}}.
\item Detailed vignette, \code{vignette("knit_print", package = "knitr")}.
\item Related \code{\link[printr:printr]{printr}} package.
}

Other Knitr printer methods: 
\code{\link{printer_tabset}()}
}
\concept{Knitr printer methods}
