% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr-printers.R
\name{printer_tabset}
\alias{printer_tabset}
\title{Print Tabsets (via \code{knit_print})}
\usage{
printer_tabset(x, options, ...)
}
\arguments{
\item{x}{list. Each element is passed to \code{knit_print} within tabs.}

\item{options, ...}{arguments required for a \code{\link[knitr:knit_print]{knit_print}}
method.}
}
\description{
Converts a list to a Quarto / RMD tabset using the names as tab headings.
}
\section{Reporting}{

These methods are to be used in Quarto / RMD reports.

For global usage on all objects of class \code{foo}, register as an S3 method,
typically in the setup chunk.\if{html}{\out{<div class="sourceCode r">}}\preformatted{library(knitr)
registerS3method("knit_print", "foo", printer_*)
}\if{html}{\out{</div>}}

For one-off usage, set the \code{render} chunk option, that is:\if{html}{\out{<div class="sourceCode markdown">}}\preformatted{```\{r\}
#| render: !expr printer_*

```
}\if{html}{\out{</div>}}
}

\seealso{
\itemize{
\item Generic documentation, \code{\link[knitr:knit_print]{knit_print}}.
\item Detailed vignette, \code{vignette("knit_print", package = "knitr")}.
\item Related \code{\link[printr:printr]{printr}} package.
}

Other Knitr printer methods: 
\code{\link{printer_equation}()}
}
\concept{Knitr printer methods}
