% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{repr_pvals}
\alias{repr_pvals}
\title{Representation of p-values
Deprecated in favour of \code{\link{format_p_values}}.}
\usage{
repr_pvals(pv, digits = 3)
}
\arguments{
\item{pv, digits}{N/A.}
}
\description{
Representation of p-values
Deprecated in favour of \code{\link{format_p_values}}.
}
