% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-testdata.R
\name{test_utils}
\alias{test_utils}
\alias{create_test_data}
\title{Test Utilities}
\usage{
create_test_data(nrows = 20, incl_na = TRUE)
}
\arguments{
\item{nrows}{integer. The number of rows that the output should have.}

\item{incl_na}{logical. When TRUE, the missing data columns are included.}
}
\value{
test data is a tibble object with all variables contained.
}
\description{
Various function to assist with unit tests.
}
\section{Test Data}{

\subsection{Explanatory Variables}{

The explanatory variables are as follows and prefixed with \code{expl_}.
\itemize{
\item \strong{rnorm}, \strong{runif}, etc.
Draws from an arbitrarily chosen distribution of the same name.
\item \strong{bin}.
Vector made up of TRUE / FALSE values, chosen at random.
\item \strong{fct}.
Factor column, uppercase letters used as levels.
\item \strong{fct_drop}.
Factor column, lowercase letters used as levels and one droppable level
with no observations.
\item \strong{na}.
Column with some values set to NA.
}
}

\subsection{Response variables:}{

The response variables are a functions of independent linear predictors.
Details are as follows and each variable name is prefixed with \code{resp_}.
\itemize{
\item \strong{lp1}, \strong{lp2}. Standard regression, \code{Y = XB}.
\item \strong{exp}. Positive values only, \code{Y = exp(XB)}.
\item \strong{exp_int}. Positive integers only, \code{Y = round(exp(XB), 0)}.
\item \strong{inv_logit}. Inverse logit transformation, \code{Y = plogis(XB, 0, 1)}.
\item \strong{na}. Standard regression with missing values mirrored in \code{expl_na}.
}
}
}

