% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testdata}
\alias{testdata}
\title{Data to be used in unit tests.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 100 rows and 16 columns.
}
\usage{
testdata
}
\description{
Haven't implemented this yet, will return a redundant string.
}
\keyword{datasets}
